Edited pinouts, added photos of 2.0 mock up.
-Edited the pinouts for the Keypad to PCB, wires were soldered
to the Keypad so these are permanent.

-Added photos of the 2.0 Housing, 2.1 Front plate, and 1.1 back plate.
I am satisifed with the overall dimensions of the control panel. I still
need to adjust the dimensions of the LCD cutout because I want it to snap
in more securely.

As you can see, the keypad dimensions were way different when I actually got the
physicaly keypad compared to the dimensions Amazon gave.

I will also change the threaded holes into thru holes because the threads serve no purpose.

There will also need to be LED knockouts on the front plate as well.