Finalized mock up photos
-Added photos of the finalized design of my mock
-up control panel box. I am satisfied with the design,
next I will work on making it look more visually pleasing.

TO DO:

-Make a more discreet fastner design, but it still has to be durable.
Will most likely remove the bolts and go with a snap design.

-Pleased with the keypad and LCD cutouts, they snap in nicely. I will
glue the perimieter of these components to keep them in the knock out.

-Will have to figure out how I want to mount the control panel while
keeping it visually pleasing and durable.

-Removed old archives folders as those designs are not needed anymore.
This is the finalized Mock up design, next I will be moving forward
with an Alpha design.